my svn update

+++ protocols/IcqOscarJ/src/capabilities.h
@@ -6,6 +6,7 @@
 // Copyright © 2001-2002 Jon Keating, Richard Hughes
 // Copyright © 2002-2004 Martin Öberg, Sam Kothari, Robert Rainwater
 // Copyright © 2004-2010 Joe Kucera
+// Copyright © 2012-2014 Miranda NG Team
 // 
 // This program is free software; you can redistribute it and/or
 // modify it under the terms of the GNU General Public License
@@ -20,17 +21,14 @@
 // You should have received a copy of the GNU General Public License
 // along with this program; if not, write to the Free Software
 // Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
-//
 // -----------------------------------------------------------------------------
 //  DESCRIPTION:
 //
 //  Contains helper functions to handle OSCAR user capabilities.
-//
 // -----------------------------------------------------------------------------
 
 #ifndef __CAPABILITIES_H
 #define __CAPABILITIES_H
-
 
 // capabilities
 typedef BYTE capstr[BINARY_CAP_SIZE];
@@ -38,8 +36,8 @@ typedef BYTE shortcapstr[BINARY_SHORT_CAP_SIZE];
 
 extern const capstr capShortCaps;
 
-capstr* MatchCapability(BYTE *buf, int bufsize, const capstr *cap, int capsize = BINARY_CAP_SIZE);
-capstr* MatchShortCapability(BYTE *buf, int bufsize, const shortcapstr *cap);
+capstr* MatchCapability(BYTE *buf, size_t bufsize, const capstr *cap, size_t capsize = BINARY_CAP_SIZE);
+capstr* MatchShortCapability(BYTE *buf, size_t bufsize, const shortcapstr *cap);
 
 
 #endif /* __CAPABILITIES_H */
